COD-50: реорганизация классов
