+++ packages/client/src/components/GameField/Game/Figure/types/index.d.ts
@@ -6,7 +6,3 @@ export interface BaseFigureProps {
   color: chipColors;
   radius: number;
 }
-
-export interface MovingFigureProps extends BaseFigureProps {
-  withoutSlot?: true;
-}
